FIX(client): Renames mispelled fn file
