+++ packages/client/src/lib/api/fetchLocations.ts
